Preserve Composer cache directory in MagentoComposerApplication
MagentoComposerApplication changes the COMPOSER_HOME environment variable.

This has a lot of unwanted side-effects, one of them is the change of the
composer cache directory which then is broken.

The fix in this commit for the isolated cache detail is to preserve the
the cache from the previous location if not already set in the
COMPOSER_CACHE environment variable.

This is done by adding a helper class that is taking care of changing the
COMPOSER environment.